requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.14.0
idna==3.6
include==0.2.2
Markdown==3.5.1
mysqlclient==2.2.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
